fix bug for markers undefined
