+++ client/src/components/Contact/Map.jsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useRef } from "react";
 import mapboxgl from "mapbox-gl";
-import handleMapViewMove from "../../helpers/handleMapViewMove";
 
 mapboxgl.accessToken =
 	"pk.eyJ1IjoibGljaGFuZzAwNCIsImEiOiJjbGZsZTdxajQwMTBkM3Fxem1xOGE4bHZhIn0.NX4YQSsBGEHQ1V_Ytk0Usg";
@@ -32,10 +31,27 @@ export default function Map({ stores, view }) {
 					.addTo(map);
 				markers.push(marker);
 			});
+		const handleMapViewMove = function () {
+			markers &&
+				markers.forEach((marker) => {
+					const popup = marker.getPopup();
+					if (!popup) return;
+					if (
+						marker.getLngLat().lat.toFixed(4) ===
+							parseFloat(view.latitude).toFixed(4) &&
+						marker.getLngLat().lng.toFixed(4) ===
+							parseFloat(view.longitude).toFixed(4)
+					) {
+						popup.addTo(map);
+					} else {
+						popup.remove();
+					}
+				});
+		};
 
 		map.on("move", handleMapViewMove);
-		handleMapViewMove(markers, view, map);
-	}, [stores.stores.stores, view, view.latitude, view.longitude]);
+		handleMapViewMove();
+	}, [stores.stores.stores, view.latitude, view.longitude]);
 
 	return (
 		<div
