+++ server/db/queries/sub_categories.js
@@ -3,7 +3,7 @@ const db = require("../connection");
 // GET ALL PRODUCTS FROM SUB_CATEGORIES NAME
 const getProductsBySubCategoryName = (subCategoryName) => {
 	const queryTemplate = `
-    SELECT make, model, image_url, description, rent_rate_in_cents, price_in_cents, products.id as id, categories.name as category_name, sub_categories.name as sub_category_name
+    SELECT make, model, image_url, description, rent_rate_in_cents, price_in_cents, products.id as id, categories.name as category_name, sub_categories.name as sub_category_name, stock_quantity
 		FROM categories
 		JOIN sub_categories ON sub_categories.category_id = categories.id
     JOIN products ON products.sub_category_id = sub_categories.id
